TT4: Task component & Delete functionality
Task.js component handles displaying information of each task. This is passes down from App -> Tasks -> Task

onDelete, in App.js, handles the deletion of a Task. Upon a deletion click, this is passed up from Task -> Tasks -> App, where App changes the state with setTasks. This deletion is based on the id of the task from the UI.
*on refresh, everything is back to normal because data is static in App.js